refactor(transformers): extract shared parquet helpers module
Move count_parquet_rows, discover_dataset_paths and to_parquet_glob (plus their discovery sub-helpers) into pts.transformers.parquet_helpers. release_metrics and dataset_metrics now import them from the shared module, removing dataset_metrics' coupling to release_metrics' private helpers. No behaviour change; release_metrics tests updated to import and monkeypatch the new location.
